refinement

+++ acceleration_d/accelerationd.c
@@ -68,10 +68,11 @@ static int poll_sensor_data(
 			buffer[i].acceleration.y,
 			buffer[i].acceleration.z);
 
-		out->x = buffer[i].acceleration.x * 100.0;
-		out->y = buffer[i].acceleration.y * 100.0;
-		out->z = buffer[i].acceleration.z * 100.0;
+		out->x = (int)(buffer[i].acceleration.x * 100.0);
+		out->y = (int)(buffer[i].acceleration.y * 100.0);
+		out->z = (int)(buffer[i].acceleration.z * 100.0);
 	}
+
 	return 0;
 }
